requirements.txt
tensorflow-gpu~=2.4.0
music21~=6.5.0
numpy~=1.19.5
matplotlib~=3.3.3
Keras~=2.4.3
sklearn
scikit-learn~=0.24.0
progress~=1.5
midi2audio~=0.1.1